fix(native): espelho de cena em dupla precisao — sombras em bandas na pista
A primeira versao expos as matrizes como Float32Array. O jogo rodou e o
fps subiu, mas as sombras da pista sairam em BANDAS: o matrixWorld do
three e double, e numa cidade (coordenadas na casa das centenas de
metros) o float32 perde digitos suficientes para o shadow map ganhar
faixas — shadow acne nascido do lado do CPU.

Matriz local, matriz de mundo e buffer de sincronizacao passam a double.
Custo: 166 KB a mais; rpMatrix ficou igual (0,205 ms). Bandas sumiram,
conferido por captura do jogo.

Regra que fica: o que espelha estado do three usa a MESMA precisao que
ele — a diferenca nao aparece como erro, aparece como artefato visual.

Medido agora: frame 38 -> 31,7 ms, fps 26,3 -> 31,5, render 21 -> 18,5.

+++ native/src/webgpu/render_bench.cpp
@@ -133,16 +133,16 @@ void benchSceneMirror(int objetos, int frames) {
   }
 
   // O buffer que o JS escreveria: só os nós que mudaram.
-  std::vector<float> sync(static_cast<size_t>(kMovingNodes) * scene::kSyncFloatsPerNode);
+  std::vector<double> sync(static_cast<size_t>(kMovingNodes) * scene::kSyncFloatsPerNode);
 
   using Relogio = std::chrono::steady_clock;
   double nanos = 0;
   long long visiveisTotal = 0;
   for (int frame = 0; frame < frames + kWarmupFrames; frame++) {
     for (int i = 0; i < kMovingNodes; i++) {
-      float* row = &sync[static_cast<size_t>(i) * scene::kSyncFloatsPerNode];
-      row[0] = static_cast<float>(i);
-      row[1] = static_cast<float>(frame) * 0.01f;
+      double* row = &sync[static_cast<size_t>(i) * scene::kSyncFloatsPerNode];
+      row[0] = static_cast<double>(i);
+      row[1] = static_cast<double>(frame) * 0.01;
       row[2] = 1.0f;
       row[3] = 0.0f;
       row[4] = 0.0f; row[5] = 0.0f; row[6] = 0.0f; row[7] = 1.0f;
